+++ 8501-Coursework/Approach2.h
@@ -4,11 +4,11 @@
 class Approach2 : public Approach {
 public:
 	virtual result run(std::vector<int>* input) override;
-	result run_experimental(std::vector<int>* input, const int* xOffset);
+	result run_experimental(std::vector<int>* input, const float* xOffset);
 protected:
 	void calculate_order_and_lead_coeff(const std::vector<int>* input, int* order, int* leadCoeff);
 	void calculate_remaining_coeffs(int* terms, const int* highorder, std::vector<int>* set);
-	void calculate_remaining_coeffs_experimental(int* terms, const int* highorder, std::vector<int>* set, const int* xOffset);
+	void calculate_remaining_coeffs_experimental(int* terms, const int* highorder, std::vector<int>* set, const float* xOffset);
 	
 	
 };
